+++ tests/Unit/Commands/DeployTest.php
@@ -3,21 +3,23 @@
  * @author Aaron Francis <aaron@hammerstone.dev>
  */
 
-namespace Hammerstone\Sidecar\Tests;
+namespace Hammerstone\Sidecar\Tests\Unit;
 
 use Hammerstone\Sidecar\Sidecar;
-use Hammerstone\Sidecar\Tests\Support\DeploymentTestFunction;
+use Hammerstone\Sidecar\Tests\Unit\Support\DeploymentTestFunction;
 
 class DeployTest extends DeploymentTest
 {
     /** @test */
     public function it_deploys_the_functions_in_the_config()
     {
+        $this->lambda->shouldReceive('functionExists')->andReturn(false);
+        $this->lambda->shouldReceive('getVersions')->andReturn([]);
+        $this->mockCreatingFunction();
+
         config()->set('sidecar.functions', [
             DeploymentTestFunction::class
         ]);
-
-        $this->mockCreatingFunction();
 
         $this->artisan('sidecar:deploy');
 
@@ -27,12 +29,14 @@ class DeployTest extends DeploymentTest
     /** @test */
     public function it_deploys_and_activates_the_functions_in_the_config()
     {
+        $this->lambda->shouldReceive('functionExists')->andReturn(false);
+        $this->lambda->shouldReceive('getVersions')->andReturn([]);
+        $this->mockCreatingFunction();
+        $this->mockActivating();
+
         config()->set('sidecar.functions', [
             DeploymentTestFunction::class
         ]);
-
-        $this->mockCreatingFunction();
-        $this->mockActivating();
 
         $this->artisan('sidecar:deploy', [
             '--activate' => true
@@ -44,6 +48,9 @@ class DeployTest extends DeploymentTest
     /** @test */
     public function it_uses_a_fake_environment()
     {
+        $this->lambda->shouldReceive('functionExists')->andReturn(false);
+        $this->lambda->shouldReceive('getVersions')->andReturn([]);
+
         config()->set('sidecar.functions', [
             DeploymentTestFunction::class
         ]);