+++ tests/Unit/DeploymentTest.php
@@ -3,7 +3,7 @@
  * @author Aaron Francis <aaron@hammerstone.dev>
  */
 
-namespace Hammerstone\Sidecar\Tests;
+namespace Hammerstone\Sidecar\Tests\Unit;
 
 use Aws\Lambda\Exception\LambdaException;
 use Hammerstone\Sidecar\Clients\LambdaClient;
@@ -12,8 +12,8 @@ use Hammerstone\Sidecar\Events\AfterFunctionsActivated;
 use Hammerstone\Sidecar\Events\AfterFunctionsDeployed;
 use Hammerstone\Sidecar\Events\BeforeFunctionsActivated;
 use Hammerstone\Sidecar\Events\BeforeFunctionsDeployed;
-use Hammerstone\Sidecar\Exceptions\ConfigurationException;
-use Hammerstone\Sidecar\Tests\Support\DeploymentTestFunction;
+use Hammerstone\Sidecar\Exceptions\NoFunctionsRegisteredException;
+use Hammerstone\Sidecar\Tests\Unit\Support\DeploymentTestFunction;
 use Illuminate\Support\Facades\Event;
 use Mockery;
 
@@ -45,8 +45,6 @@ class DeploymentTest extends BaseTest
 
     public function mockCreatingFunction()
     {
-        $this->mockFunctionNotExisting();
-
         $this->lambda->shouldReceive('createFunction')->once()->with([
             'FunctionName' => 'test-FunctionName',
             'Runtime' => 'test-Runtime',
@@ -67,69 +65,23 @@ class DeploymentTest extends BaseTest
         $this->lambda->shouldNotReceive('updateFunctionCode');
     }
 
-    public function mockListingVersions()
-    {
-        $this->lambda->shouldReceive('listVersionsByFunction')
-            ->once()
-            ->with([
-                'FunctionName' => 'test-FunctionName',
-                'MaxItems' => 100,
-                'Marker' => null
-            ])
-            ->andReturn([
-                'Versions' => [[
-                    'FunctionName' => 'test-FunctionName',
-                    'Version' => '10',
-                ], [
-                    'FunctionName' => 'test-FunctionName',
-                    'Version' => '11',
-                ], [
-                    'FunctionName' => 'test-FunctionName',
-                    'Version' => '12',
-                ]]
-            ]);
-    }
-
-    public function mockUpdating()
-    {
-        $this->lambda->shouldReceive('getFunction')->andReturn([
-            'Configuration' => [
-                'FunctionName' => 'test-FunctionName',
-                'Description' => 'test-Description',
-            ]
-        ]);
-
-        $this->lambda->shouldReceive('updateFunctionConfiguration')->with([
-            'FunctionName' => 'test-FunctionName',
-            'Role' => 'test-Role',
-            'Handler' => 'test-Handler',
-            'Description' => 'test-Description [5000a525]',
-            'Timeout' => 'test-Timeout',
-            'MemorySize' => 'test-MemorySize'
-        ]);
-
-        $this->lambda->shouldReceive('updateFunctionCode')->with([
-            'FunctionName' => 'test-FunctionName',
-            'Publish' => 'test-Publish',
-            'S3Bucket' => 'test-bucket',
-            'S3Key' => 'test-key'
-        ]);
-    }
-
     public function mockActivating()
     {
-        $this->mockListingVersions();
+        $this->lambda->shouldReceive('getLatestVersion')
+            ->once()
+            ->withArgs(function ($function) {
+                return $function instanceof DeploymentTestFunction;
+            })
+            ->andReturn('10');
 
-        $this->lambda->shouldReceive('deleteAlias')->once()->with([
-            'FunctionName' => 'test-FunctionName',
-            'Name' => 'active',
-        ]);
-
-        $this->lambda->shouldReceive('createAlias')->once()->with([
-            'FunctionName' => 'test-FunctionName',
-            'FunctionVersion' => '12',
-            'Name' => 'active',
-        ]);
+        $this->lambda->shouldReceive('aliasVersion')
+            ->once()
+            ->withArgs(function ($function, $alias, $version) {
+                return $function instanceof DeploymentTestFunction
+                    && $alias === 'active'
+                    && $version === '10';
+            })
+            ->andReturn(LambdaClient::CREATED);
     }
 
     public function assertEvents($deployed = true, $activated = true)
@@ -154,6 +106,8 @@ class DeploymentTest extends BaseTest
     /** @test */
     public function it_deploys_a_function_that_doesnt_exist()
     {
+        $this->lambda->shouldReceive('functionExists')->andReturn(false);
+        $this->lambda->shouldReceive('getVersions')->andReturn([]);
         $this->mockCreatingFunction();
 
         DeploymentTestFunction::deploy($activate = false);
@@ -164,6 +118,8 @@ class DeploymentTest extends BaseTest
     /** @test */
     public function it_deploys_a_function_that_doesnt_exist_from_the_deployment_class()
     {
+        $this->lambda->shouldReceive('functionExists')->andReturn(false);
+        $this->lambda->shouldReceive('getVersions')->andReturn([]);
         $this->mockCreatingFunction();
 
         Deployment::make(DeploymentTestFunction::class)->deploy($activate = false);
@@ -174,6 +130,8 @@ class DeploymentTest extends BaseTest
     /** @test */
     public function it_deploys_an_array_of_functions()
     {
+        $this->lambda->shouldReceive('functionExists')->andReturn(false);
+        $this->lambda->shouldReceive('getVersions')->andReturn([]);
         $this->mockCreatingFunction();
 
         Deployment::make([DeploymentTestFunction::class])->deploy($activate = false);
@@ -184,11 +142,13 @@ class DeploymentTest extends BaseTest
     /** @test */
     public function it_deploys_the_functions_in_the_config()
     {
+        $this->lambda->shouldReceive('functionExists')->andReturn(false);
+        $this->lambda->shouldReceive('getVersions')->andReturn([]);
+        $this->mockCreatingFunction();
+
         config()->set('sidecar.functions', [
             DeploymentTestFunction::class
         ]);
-
-        $this->mockCreatingFunction();
 
         Deployment::make()->deploy($activate = false);
 
@@ -198,6 +158,8 @@ class DeploymentTest extends BaseTest
     /** @test */
     public function it_deploys_and_activates_a_function_that_doesnt_exist()
     {
+        $this->lambda->shouldReceive('functionExists')->andReturn(false);
+        $this->lambda->shouldReceive('getVersions')->andReturn([]);
         $this->mockCreatingFunction();
         $this->mockActivating();
 
@@ -209,7 +171,11 @@ class DeploymentTest extends BaseTest
     /** @test */
     public function it_updates_an_existing_function()
     {
-        $this->mockUpdating();
+        $this->lambda->shouldReceive('functionExists')->andReturn(true);
+        $this->lambda->shouldReceive('getVersions')->andReturn([]);
+        $this->lambda->shouldReceive('updateExistingFunction')->once()->withArgs(function ($function) {
+            return $function instanceof DeploymentTestFunction;
+        });
 
         DeploymentTestFunction::deploy($activate = false);
 
@@ -219,28 +185,12 @@ class DeploymentTest extends BaseTest
     /** @test */
     public function it_updates_and_activates_an_existing_function()
     {
-        $this->mockUpdating();
+        $this->lambda->shouldReceive('functionExists')->andReturn(true);
+        $this->lambda->shouldReceive('getVersions')->andReturn([]);
+        $this->lambda->shouldReceive('updateExistingFunction')->once()->withArgs(function ($function) {
+            return $function instanceof DeploymentTestFunction;
+        });
         $this->mockActivating();
-
-        DeploymentTestFunction::deploy($activate = true);
-
-        $this->assertEvents($deployed = true, $activated = true);
-    }
-
-    /** @test */
-    public function it_doesnt_update_if_nothing_has_changed()
-    {
-        $this->mockActivating();
-
-        $this->lambda->shouldReceive('getFunction')->twice()->andReturn([
-            'Configuration' => [
-                'FunctionName' => 'test-FunctionName',
-                'Description' => 'test-Description [5000a525]',
-            ]
-        ]);
-
-        $this->lambda->shouldNotReceive('updateFunctionConfiguration');
-        $this->lambda->shouldNotReceive('updateFunctionCode');
 
         DeploymentTestFunction::deploy($activate = true);
 
@@ -254,8 +204,7 @@ class DeploymentTest extends BaseTest
 
         ]);
 
-        $this->expectException(ConfigurationException::class);
-        $this->expectExceptionMessage('Cannot deploy, no Sidecar functions have been configured');
+        $this->expectException(NoFunctionsRegisteredException::class);
 
         Deployment::make()->deploy();
     }