+++ tests/Unit/ExecuteMultipleTest.php
@@ -3,7 +3,7 @@
  * @author Aaron Francis <aaron@hammerstone.dev>
  */
 
-namespace Hammerstone\Sidecar\Tests;
+namespace Hammerstone\Sidecar\Tests\Unit;
 
 use Aws\Result;
 use GuzzleHttp\Promise\PromiseInterface;
@@ -12,7 +12,7 @@ use Hammerstone\Sidecar\Events\AfterFunctionExecuted;
 use Hammerstone\Sidecar\Events\BeforeFunctionExecuted;
 use Hammerstone\Sidecar\Results\PendingResult;
 use Hammerstone\Sidecar\Sidecar;
-use Hammerstone\Sidecar\Tests\Support\EmptyTestFunction;
+use Hammerstone\Sidecar\Tests\Unit\Support\EmptyTestFunction;
 use Illuminate\Support\Facades\Event;
 use Mockery;
 
@@ -28,7 +28,7 @@ class ExecuteMultipleTest extends BaseTest
     protected function expectedArgs($args = [])
     {
         return array_merge([
-            'FunctionName' => 'SC-Laravel-testing-stone-Sidecar-Tests-Support-EmptyTestFunction:active',
+            'FunctionName' => 'SC-Laravel-testing--Sidecar-Tests-Unit-Support-EmptyTestFunction:active',
             'InvocationType' => 'RequestResponse',
             'LogType' => 'Tail',
             'Payload' => '[]'